requirements.txt
pydantic>=2.6
httpx>=0.24
python-docx>=1.1
python-pptx>=0.6.21
pypdf>=4.0
pyyaml>=6.0
fireworks-ai>=0.15
